+++ .dockerignore
@@ -1,0 +1,24 @@
+# 노드 모듈 제외
+node_modules
+
+# NestJS 빌드 폴더 제외
+dist
+
+# 환경 설정 파일 제외
+.env
+
+# Git 관련 파일/디렉토리 제외
+.git
+.gitignore
+
+# npm 디버그 로그
+npm-debug.log
+
+# 에디터 설정 파일 제외
+.vscode
+*.swp
+*.swo
+
+# 운영체제 생성 파일 제외
+.DS_Store
+Thumbs.db
