--- configs/typeOrmModuleOption.ts
@@ -1,24 +1,0 @@
-import { ConfigModule, ConfigService } from "@nestjs/config";
-import {
-  TypeOrmModuleAsyncOptions,
-  TypeOrmModuleOptions,
-} from "@nestjs/typeorm";
-import { SnakeNamingStrategy } from "typeorm-naming-strategies";
-export const typeOrmModuleOptions: TypeOrmModuleAsyncOptions = {
-  imports: [ConfigModule],
-  inject: [ConfigService],
-  useFactory: async (
-    configService: ConfigService,
-  ): Promise<TypeOrmModuleOptions> => ({
-    namingStrategy: new SnakeNamingStrategy(),
-    type: "mysql",
-    host: configService.get<string>("DB_HOST"),
-    port: configService.get<number>("DB_PORT"),
-    username: configService.get<string>("DB_USERNAME"),
-    password: configService.get<string>("DB_PASSWORD"),
-    database: configService.get<string>("DB_NAME"),
-    synchronize: configService.get<boolean>("DB_SYNC"),
-    autoLoadEntities: true,
-    logging: true,
-  }),
-};
