+++ src/app.module.ts
@@ -1,15 +1,16 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { ConfigModule } from '@nestjs/config';
-import { configModuleValidationSchema } from 'configs/env_validation';
+import { configModuleValidationSchema } from 'src/configs/env_validation';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { typeOrmModuleOptions } from 'configs/typeOrmModuleOption';
+import { typeOrmModuleOptions } from 'src/configs/typeOrmModuleOption';
 import { OrdersModule } from './orders/orders.module';
 import { PaymentsModule } from './payments/payments.module';
 import { UsersModule } from './user/users.module';
-import { GoodsModule } from "./goods/goods.module";
-import { AuthModule } from "./auth/auth.module";
+import { GoodsModule } from './goods/goods.module';
+import { AuthModule } from './auth/auth.module';
 import { RedisModule } from './redis/redis.module';
+import { BoardsModule } from './boards/boards.module';
 
 @Module({
   imports: [
@@ -25,6 +26,7 @@ import { RedisModule } from './redis/redis.module';
     UsersModule,
     GoodsModule,
     RedisModule,
+    BoardsModule,
   ],
   controllers: [AppController],
   providers: [],
