+++ src/configs/env_validation.ts
@@ -1,0 +1,16 @@
+import Joi from 'joi';
+
+export const configModuleValidationSchema = Joi.object({
+  SERVER_PORT: Joi.number().default(3000),
+  DB_HOST: Joi.string().default('localhost'),
+  DB_PORT: Joi.number().default(3306),
+  DB_USERNAME: Joi.string().required(),
+  DB_PASSWORD: Joi.string().required(),
+  DB_NAME: Joi.string().required(),
+  DB_SYNC: Joi.boolean().default(false),
+  PASSWORD_HASH_ROUND: Joi.number().default(10),
+
+  //JWT 서명 용
+  JWT_ACCESS_TOKEN_SECRET: Joi.string().required(),
+  JWT_REFRESH_TOKEN_SECRET: Joi.string().required(),
+});
